Split document head selectors (#43755)
* Create new folder for document-head selectors

* Add `package.json`

* Split out `getDocumentHeadTitle`

* Split out `getDocumentHeadUnreadCount`

* Split out `getDocumentHeadCappedUnreadCount`

* Split out `getDocumentHeadFormattedTitle`

* Split out `getDocumentHeadMeta`

* Split out `getDocumentHeadLink`

* Sort exports alphabetically

* Get document head selectors directly
