+++ client/state/document-head/selectors/get-document-head-meta.js
@@ -1,0 +1,10 @@
+/**
+ * Returns an array of document meta objects as set by the DocumentHead
+ * component or setDocumentHeadMeta action.
+ *
+ * @param  {object}  state  Global state tree
+ * @returns {object[]}       Array of meta objects
+ */
+export function getDocumentHeadMeta( state ) {
+	return state.documentHead.meta;
+}
