+++ client/state/document-head/selectors/get-document-head-title.js
@@ -1,0 +1,10 @@
+/**
+ * Returns the document title as set by the DocumentHead component or setTitle
+ * action.
+ *
+ * @param  {object}  state  Global state tree
+ * @returns {?string}        Document title
+ */
+export function getDocumentHeadTitle( state ) {
+	return state.documentHead.title;
+}
