+++ client/state/document-head/selectors/index.js
@@ -1,0 +1,6 @@
+export { getDocumentHeadCappedUnreadCount } from 'state/document-head/selectors/get-document-head-capped-unread-count';
+export { getDocumentHeadFormattedTitle } from 'state/document-head/selectors/get-document-head-formatted-title';
+export { getDocumentHeadLink } from 'state/document-head/selectors/get-document-head-link';
+export { getDocumentHeadMeta } from 'state/document-head/selectors/get-document-head-meta';
+export { getDocumentHeadTitle } from 'state/document-head/selectors/get-document-head-title';
+export { getDocumentHeadUnreadCount } from 'state/document-head/selectors/get-document-head-unread-count';
