+++ client/state/document-head/selectors/package.json
@@ -1,0 +1,3 @@
+{
+	"sideEffects": false
+}
